requirements.txt
tensorflow>=2.12.0
streamlit>=1.22.0
numpy>=1.23.5
pillow>=9.5.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
pandas>=2.0.0 